requirements.txt
Flask==1.1.2
boto3==1.16.43
python-decouple==3.3
torch==1.7.1
torchvision==0.8.2